✨ Keybinds are now embedded into WindowDirection
